+++ hardev/web/prettier.config.mjs
@@ -1,0 +1,7 @@
+/** @type {import("prettier").Config} */
+export default {
+  semi: false,
+  singleQuote: true,
+  trailingComma: 'es5',
+  printWidth: 100,
+}